Split resolution.rs into multiple files
